fix: add z-index and move QToastr component to QToastr folder
add z-index to Toastr.scss and move QToastr and its styles to QToastr folder
