+++ packages/toastr/lib/components/Toastr/styles/QToastr.scss
