+++ packages/toastr/src/react-component-node-modules/components/QToastr/QToastr.md
@@ -1,0 +1,10 @@
+```js
+<div>
+  <QToastr toastrStyle="qts">已複製到剪貼簿。</QToastr>
+  <QToastr toastrStyle="photo-st">PhotoStation</QToastr>
+  <QToastr toastrStyle="music-st">MusicStation</QToastr>
+  <QToastr toastrStyle="file-st">FileStation</QToastr>
+  <QToastr toastrStyle="download-st">DownloadStation</QToastr>
+  <QToastr toastrStyle="video-st">VideoStation</QToastr>
+</div>
+```
