+++ packages/toastr/src/react-component-node-modules/components/QToastr/styles/QToastr.scss
@@ -1,0 +1,41 @@
+@import 'colors';
+@import 'mixins';
+
+.toastr {
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  justify-content: center;
+  min-width: 64px;
+  padding: 24px;
+  margin-top: 24px;
+  font-family: Roboto, Helvetica, Arial, Sans-serif, Verdana, Tahoma;
+  font-size: 16px;
+  line-height: 20px;
+  color: $toastr-font-color;
+  text-align: center;
+}
+
+.qts {
+  background-color: $qts-default-color;
+}
+
+.photo-st {
+  background-color: $photo-station-color;
+}
+
+.music-st {
+  background-color: $music-station-color;
+}
+
+.file-st {
+  background-color: $file-station-color;
+}
+
+.download-st {
+  background-color: $download-station-color;
+}
+
+.video-st {
+  background-color: $video-station-color;
+}
