+++ packages/toastr/src/react-component-node-modules/components/QToastr/styles/_mixins.scss
