+++ packages/toastr/src/react-component-node-modules/components/Toastr/styles/Toastr.scss
@@ -1,5 +1,6 @@
 .view-page {
   position: fixed;
+  z-index: 99;
   display: flex;
   align-items: flex-start;
   justify-content: center;
