+++ packages/toastr/src/react-component-node-modules/index.js
@@ -1,2 +1,2 @@
-export { default as QToastr } from './components/Toastr/QToastr';
+export { default as QToastr } from './components/QToastr/QToastr';
 export { default as Toastr } from './components/Toastr/Toastr'
